Attempt to Fix Calibration via Menu
Not sure why, but for some reason the z-height is not being set in
eeprom. I'm trying to find a fix.

+++ src/ArduinoDUE/Repetier/Commands.cpp
@@ -421,7 +421,7 @@ void microstepInit()
 
 
 #if FEATURE_AUTOLEVEL
-void Commands::AutoLevel(byte S, bool start)
+void Commands::AutoLevel(int S,bool start)
       {
         switch (start) 
         {
